backend/chatbot/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
qdrant-client
openai
python-dotenv
langchain
